rename graph_parameters
