+++ cpp/benchmarks/secir_ageres_setups.h
@@ -21,7 +21,7 @@
 #define SECIR_AGERES_SETUPS_H_
 
 #include "memilio/compartments/simulation.h"
-#include "memilio/mobility/graph_parameters.h"
+#include "memilio/mobility/meta_mobility_instant.h"
 #include "memilio/compartments/parameter_studies.h"
 #include "models/ode_secir/model.h"
 #include "models/ode_secir/parameter_space.h"
