+++ cpp/examples/graph.cpp
@@ -20,7 +20,7 @@
 #include "ode_seir/model.h"
 #include "ode_seir/infection_state.h"
 #include "ode_seir/parameters.h"
-#include "memilio/mobility/graph_parameters.h"
+#include "memilio/mobility/meta_mobility_instant.h"
 #include "memilio/compartments/simulation.h"
 
 #include <iostream>
