+++ cpp/memilio/compartments/parameter_studies.h
@@ -21,7 +21,7 @@
 #define PARAMETER_STUDIES_H
 
 #include "memilio/utils/time_series.h"
-#include "memilio/mobility/graph_parameters.h"
+#include "memilio/mobility/meta_mobility_instant.h"
 #include "memilio/compartments/simulation.h"
 
 #include <cmath>
