+++ cpp/memilio/io/mobility_io.h
@@ -24,7 +24,7 @@
 #include "memilio/math/eigen.h"
 #include "memilio/io/json_serializer.h"
 #include "memilio/mobility/graph.h"
-#include "memilio/mobility/graph_parameters.h"
+#include "memilio/mobility/meta_mobility_instant.h"
 #include <iostream>
 #include <fstream>
 #include <string>
