+++ cpp/tests/test_mobility.cpp
@@ -19,7 +19,7 @@
 */
 #define _USE_MATH_DEFINES
 
-#include "memilio/mobility/graph_parameters.h"
+#include "memilio/mobility/meta_mobility_instant.h"
 #include "ode_seir/model.h"
 #include "ode_seir/infection_state.h"
 #include "ode_seir/parameters.h"
