+++ pycode/memilio-simulation/CMakeLists.txt
@@ -51,7 +51,7 @@ pybind11_add_module(_simulation MODULE
     memilio/simulation/simulation.cpp
     memilio/simulation/epidemiology/damping_sampling.cpp
     memilio/simulation/epidemiology/uncertain_matrix.cpp
-    memilio/simulation/mobility/graph_parameters.cpp
+    memilio/simulation/mobility/meta_mobility_instant.cpp
     memilio/simulation/utils/date.cpp
     memilio/simulation/utils/logging.cpp
     memilio/simulation/utils/time_series.cpp
